Refactor utils.py to reflect base changes done that have been merged into this branch from develop.
